I have reorganized the labeling layout into three columns to improve UI stability. Here are the updates I’ve made:
- Moved the image and action bar to the first column.
- Extracted the observation tally (including animals, checkboxes, and comments) into a new component in the middle column.
- Kept the species search in the third column.
- Implemented a fixed image frame and scroll areas to prevent layout shifts.
- Ensured the confirmation button remains visible without requiring the user to scroll.
- Optimized the layout to be responsive, stacking the columns for mobile devices.
