+++ wildmile/components/cameratrap/ContextCamera.js
@@ -24,6 +24,13 @@ const [CounterProvider, useCounter] = createCtx(0);
 const [ImageProvider, useImage] = createCtx(null);
 const [SelectionProvider, useSelection] = createCtx([]);
 const [RecentSpeciesProvider, useRecentSpecies] = createCtx([]);
+const [AnimalCountsProvider, useAnimalCounts] = createCtx({});
+const [ObservationStateProvider, useObservationState] = createCtx({
+  humanPresent: false,
+  vehiclePresent: false,
+  noAnimalsVisible: false,
+  comment: "",
+});
 
 // Combined provider
 function IdentificationProvider({ children }) {
@@ -31,7 +38,11 @@ function IdentificationProvider({ children }) {
     <CounterProvider>
       <ImageProvider>
         <SelectionProvider>
-          <RecentSpeciesProvider>{children}</RecentSpeciesProvider>
+          <RecentSpeciesProvider>
+            <AnimalCountsProvider>
+              <ObservationStateProvider>{children}</ObservationStateProvider>
+            </AnimalCountsProvider>
+          </RecentSpeciesProvider>
         </SelectionProvider>
       </ImageProvider>
     </CounterProvider>
@@ -44,4 +55,6 @@ export {
   useImage,
   useSelection,
   useRecentSpecies,
+  useAnimalCounts,
+  useObservationState,
 };
