+++ wildmile/components/cameratrap/ImageAnnotation.js
@@ -3,44 +3,32 @@
 import React, { useState, useEffect } from "react";
 import {
   Card,
-  CardSection,
   Image,
   Text,
   Button,
-  NumberInput,
   Group,
-  Stack,
   Badge,
-  Checkbox,
-  TextInput,
   ActionIcon,
   Modal,
   Grid,
   GridCol,
   Indicator,
-  Flex,
   Tooltip,
-  ScrollArea,
   Box,
 } from "@mantine/core";
 import {
   IconHeartPlus,
   IconHeart,
-  IconHeartFilled,
   IconSend,
   IconMaximize,
   IconLink,
-  IconX,
   IconPhotoSearch,
-  IconZoomQuestion,
   IconMoodWrrr,
   IconFocus2,
   IconPlayerPlay,
 } from "@tabler/icons-react";
 import { TransformWrapper, TransformComponent } from "react-zoom-pan-pinch";
 import { useImage, useSelection, useRecentSpecies } from "./ContextCamera";
-import checkboxClasses from "styles/checkbox.module.css";
-import styles from "styles/animalSelection.module.css";
 import { ObservationHistoryPopover } from "./ObservationHistory";
 import { SpeciesConsensusBadges } from "./SpeciesConsensusBadges";
 
@@ -48,15 +36,9 @@ export function ImageAnnotation({ fetchNextImage, filters }) {
   const [currentImage, setCurrentImage] = useImage();
   const [selection, setSelection] = useSelection();
   const [recentSpecies, setRecentSpecies] = useRecentSpecies();
-  const [animalCounts, setAnimalCounts] = useState({});
-  const [noAnimalsVisible, setNoAnimalsVisible] = useState(false);
   const [isSaving, setIsSaving] = useState(false);
-  const [comment, setComment] = useState("");
-  const [comments, setComments] = useState([]);
   const [isFavorite, setIsFavorite] = useState(false);
   const [enlargedImage, setEnlargedImage] = useState(false);
-  const [humanPresent, setHumanPresent] = useState(false);
-  const [vehiclePresent, setVehiclePresent] = useState(false);
   const [needsReview, setNeedsReview] = useState(false);
   const [flagged, setFlagged] = useState(false);
   const [showAIBoxes, setShowAIBoxes] = useState(false);
@@ -64,159 +46,11 @@ export function ImageAnnotation({ fetchNextImage, filters }) {
 
   useEffect(() => {
     if (currentImage) {
-      setComments(currentImage.mediaComments || []);
       setIsFavorite(currentImage.favorite || false);
       setNeedsReview(currentImage.needsReview || false);
       setFlagged(currentImage.flagged || false);
     }
   }, [currentImage]);
-
-  const handleCountChange = (id, value) => {
-    setAnimalCounts((prev) => ({ ...prev, [id]: value }));
-  };
-
-  const handleNoAnimalsClick = async () => {
-    await handleSaveObservations({ forceNoAnimals: true });
-  };
-
-  const handleSaveObservations = async ({ forceNoAnimals = false } = {}) => {
-    if (!currentImage) return;
-
-    setIsSaving(true);
-    if (comment.trim()) {
-      await handleAddComment();
-    }
-
-    let observations = [];
-
-    if (
-      (forceNoAnimals || noAnimalsVisible) &&
-      !humanPresent &&
-      !vehiclePresent
-    ) {
-      observations.push({
-        mediaId: currentImage.mediaID,
-        mediaInfo: {
-          md5: currentImage.mediaID,
-          imageHash: currentImage.imageHash,
-        },
-        eventStart: currentImage.timestamp,
-        eventEnd: currentImage.timestamp,
-        observationLevel: "media",
-        observationType: "blank",
-      });
-    } else {
-      if (selection.length > 0) {
-        observations = selection.map((animal) => ({
-          mediaId: currentImage.mediaID,
-          mediaInfo: {
-            md5: currentImage.mediaID,
-            imageHash: currentImage.imageHash,
-          },
-          taxonId: animal.id,
-          scientificName: animal.name,
-          commonName: animal.preferred_common_name || animal.name,
-          count: animalCounts[animal.id] || 1,
-          eventStart: currentImage.timestamp,
-          eventEnd: currentImage.timestamp,
-          observationLevel: "media",
-          observationType: "animal",
-        }));
-      }
-
-      if (humanPresent) {
-        observations.push({
-          mediaId: currentImage.mediaID,
-          mediaInfo: {
-            md5: currentImage.mediaID,
-            imageHash: currentImage.imageHash,
-          },
-          eventStart: currentImage.timestamp,
-          eventEnd: currentImage.timestamp,
-          observationLevel: "media",
-          observationType: "human",
-        });
-      }
-
-      if (vehiclePresent) {
-        observations.push({
-          mediaId: currentImage.mediaID,
-          mediaInfo: {
-            md5: currentImage.mediaID,
-            imageHash: currentImage.imageHash,
-          },
-          eventStart: currentImage.timestamp,
-          eventEnd: currentImage.timestamp,
-          observationLevel: "media",
-          observationType: "vehicle",
-        });
-      }
-    }
-
-    try {
-      const response = await fetch("/api/cameratrap/saveObservations", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(observations),
-      });
-
-      fetchNextImage();
-      if (response.ok) {
-        const data = await response.json();
-        if (data.savedSpecies?.length) {
-          setRecentSpecies((prev) => {
-            const existingNames = new Set(
-              prev.map((s) => s.name?.toLowerCase())
-            );
-            const newEntries = selection
-              .filter(
-                (s) =>
-                  data.savedSpecies.includes(s.name) &&
-                  !existingNames.has(s.name?.toLowerCase())
-              )
-              .map((s) => ({
-                ...s,
-                name: s.name,
-                preferred_common_name: s.preferred_common_name,
-              }));
-            if (!newEntries.length) return prev;
-            return [...newEntries, ...prev].slice(0, 12);
-          });
-        }
-        setNoAnimalsVisible(false);
-      } else {
-        alert("Failed to save observations. Make sure you're logged in");
-      }
-    } catch (error) {
-      console.error("Error saving observations:", error);
-      alert("Error saving observations");
-    } finally {
-      setIsSaving(false);
-    }
-  };
-
-  const handleAddComment = async () => {
-    if (!comment.trim()) return;
-
-    try {
-      const response = await fetch("/api/cameratrap/addComment", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ mediaId: currentImage.mediaID, comment }),
-      });
-
-      if (response.ok) {
-        const newComment = await response.json();
-        setComments([...comments, newComment]);
-        setComment("");
-      } else {
-        alert("Failed to add comment");
-      }
-    } catch (error) {
-      console.error("Error adding comment:", error);
-      alert("Error adding comment");
-    }
-  };
 
   const handleToggleFavorite = async () => {
     try {
@@ -283,22 +117,13 @@ export function ImageAnnotation({ fetchNextImage, filters }) {
     setEnlargedImage(!enlargedImage);
   };
 
-  const handleRemoveAnimal = (animalId) => {
-    setSelection((prev) => prev.filter((animal) => animal.id !== animalId));
-    setAnimalCounts((prev) => {
-      const newCounts = { ...prev };
-      delete newCounts[animalId];
-      return newCounts;
-    });
-  };
-
   if (!currentImage) {
     return <Text>No image selected</Text>;
   }
-  console.log(currentImage);
+
   return (
     <>
-      <Card shadow="sm" radius="md" withBorder style={{ height: "100%" }}>
+      <Card shadow="sm" radius="md" withBorder>
         <Box
           style={{
             position: "relative",
@@ -310,15 +135,11 @@ export function ImageAnnotation({ fetchNextImage, filters }) {
         >
           <TransformWrapper
             defaultScale={1}
-            wheel={{ step: 0.4 }} // how fast you zoom with the mouse wheel
-            pinch={{ step: 0.2 }} // how fast you zoom with pinch gesture
-            // doubleClick={{ disabled: true }} // optional: disable double-click zoom
+            wheel={{ step: 0.4 }}
+            pinch={{ step: 0.2 }}
           >
             <TransformComponent
-              wrapperStyle={{
-                width: "100%",
-                height: "100%",
-              }}
+              wrapperStyle={{ width: "100%", height: "100%" }}
               contentStyle={{
                 width: "100%",
                 height: "100%",
@@ -381,22 +202,10 @@ export function ImageAnnotation({ fetchNextImage, filters }) {
             <IconMaximize size={24} />
           </ActionIcon>
         </Box>
-        <Grid>
-          <GridCol span={12}>
-            <Group gap="xl">
-              <Text mt="xs" size="xs" style={{ fontFamily: "monospace" }}>
-                Image Timestamp:{" "}
-                {new Date(currentImage.timestamp).toLocaleString("en-US", {
-                  timeZone: "UTC",
-                })}
-              </Text>
-              <Text mt="xs" size="xs" style={{ fontFamily: "monospace" }}>
-                Media ID: {currentImage.mediaID}
-              </Text>
-            </Group>
-          </GridCol>
-          <GridCol span={{ base: 12, md: 12, lg: 6 }}>
-            <Group position="apart" mt="md">
+
+        <Stack gap="xs" mt="md">
+          <Group position="apart">
+            <Group gap="xs">
               <ActionIcon
                 variant="outline"
                 onClick={() => {
@@ -459,7 +268,7 @@ export function ImageAnnotation({ fetchNextImage, filters }) {
               >
                 <ActionIcon
                   onClick={handleToggleFavorite}
-                  color={isFavorite ? "red" : "red"}
+                  color="red"
                   variant={isFavorite ? "filled" : "outline"}
                 >
                   {isFavorite ? (
@@ -470,141 +279,26 @@ export function ImageAnnotation({ fetchNextImage, filters }) {
                 </ActionIcon>
               </Indicator>
             </Group>
-            <Group mt="xs">
-              <TextInput
-                placeholder="Add a comment..."
-                value={comment}
-                onChange={(event) => setComment(event.currentTarget.value)}
-                style={{ flex: 1 }}
-              />
-              <ActionIcon onClick={handleAddComment} disabled={!comment.trim()}>
-                <IconSend size={24} />
-              </ActionIcon>
-            </Group>
-            <ScrollArea h={100} offsetScrollbars mt="md">
-              <Stack gap="xs">
-                {comments.map((comment, index) => (
-                  <Text key={index} size="sm">
-                    <strong>{comment.author.name}:</strong> {comment.text}
-                  </Text>
-                ))}
-              </Stack>
-            </ScrollArea>
-          </GridCol>
-          <GridCol span={{ base: 12, md: 12, lg: 6 }}>
-            {!noAnimalsVisible && (
-              <ScrollArea
-                h={200}
-                offsetScrollbars
-                scrollbarSize={4}
-                mt="md"
-              >
-                <Flex direction="column" gap="xs">
-                  {selection.map((animal) => (
-                    <div key={animal.id} className={styles.selectionContainer}>
-                      <div className={styles.selectionContent}>
-                        <Text className={styles.speciesName}>
-                          {animal.preferred_common_name || animal.name}
-                        </Text>
-                        <div className={styles.controls}>
-                          <NumberInput
-                            value={animalCounts[animal.id] || 1}
-                            onChange={(value) =>
-                              handleCountChange(animal.id, value)
-                            }
-                            min={1}
-                            max={100}
-                            style={{ width: 80 }}
-                          />
-                          <ActionIcon
-                            color="red"
-                            variant="subtle"
-                            onClick={() => handleRemoveAnimal(animal.id)}
-                            className={styles.removeButton}
-                          >
-                            <IconX size={16} />
-                          </ActionIcon>
-                        </div>
-                      </div>
-                    </div>
-                  ))}
-                </Flex>
-              </ScrollArea>
-            )}
-            <Group mt="xs" grow wrap="nowrap">
-              <Checkbox
-                classNames={checkboxClasses}
-                label="Human Present"
-                checked={humanPresent}
-                onChange={(event) =>
-                  setHumanPresent(event.currentTarget.checked)
-                }
-                wrapperProps={{
-                  onClick: () => setHumanPresent((c) => !c),
-                }}
-              />
-              <Checkbox
-                classNames={checkboxClasses}
-                label="Vehicle Present"
-                checked={vehiclePresent}
-                onChange={(event) =>
-                  setVehiclePresent(event.currentTarget.checked)
-                }
-                wrapperProps={{
-                  onClick: () => setVehiclePresent((c) => !c),
-                }}
-              />
-            </Group>
-          </GridCol>
-          <GridCol span={12}>
-            {noAnimalsVisible ||
-            selection.length > 0 ||
-            humanPresent ||
-            vehiclePresent ? (
-              <Button
-                color="blue"
-                fullWidth
-                mt="md"
-                radius="md"
-                onClick={handleSaveObservations}
-                loading={isSaving}
-              >
-                {isSaving ? "Saving..." : "Save Observations"}
-              </Button>
-            ) : (
-              <Button
-                color="blue"
-                variant="outline"
-                fullWidth
-                mt="md"
-                radius="md"
-                onClick={handleNoAnimalsClick}
-                loading={isSaving}
-              >
-                No Animals Visible
-              </Button>
-            )}
-            {selection.length === 0 &&
-              !noAnimalsVisible &&
-              !humanPresent &&
-              !vehiclePresent && (
-                <Text color="dimmed" ta="center" mt="md">
-                  Select animals from the search results to add observations,
-                  mark as "No Animals Visible", or indicate human/vehicle
-                  presence
-                </Text>
-              )}
-          </GridCol>
-          <GridCol span={12}>
-            <Group mt="md">
-              <SpeciesConsensusBadges
-                speciesConsensus={currentImage.speciesConsensus}
-              />
-              <ObservationHistoryPopover mediaID={currentImage.mediaID} />
-            </Group>
-          </GridCol>
-        </Grid>
+          </Group>
+
+          <Group gap="xl">
+            <Text size="xs" style={{ fontFamily: "monospace" }}>
+              Time: {new Date(currentImage.timestamp).toLocaleString("en-US", { timeZone: "UTC" })}
+            </Text>
+            <Text size="xs" style={{ fontFamily: "monospace" }}>
+              ID: {currentImage.mediaID}
+            </Text>
+          </Group>
+
+          <Group mt="xs">
+            <SpeciesConsensusBadges
+              speciesConsensus={currentImage.speciesConsensus}
+            />
+            <ObservationHistoryPopover mediaID={currentImage.mediaID} />
+          </Group>
+        </Stack>
       </Card>
+
       <Modal
         opened={enlargedImage}
         onClose={() => setEnlargedImage(false)}
@@ -615,29 +309,13 @@ export function ImageAnnotation({ fetchNextImage, filters }) {
           modal: { maxWidth: "100%" },
         }}
       >
-        <div
-          style={{
-            // width: "90vw",
-            // height: "90vh",
-            display: "flex",
-            // justifyContent: "center",
-            alignItems: "center",
-            backgroundColor: "black",
-          }}
-        >
-          <TransformWrapper
-            defaultScale={1}
-            wheel={{ step: 0.4 }} // how fast you zoom with the mouse wheel
-            pinch={{ step: 0.2 }} // how fast you zoom with pinch gesture
-            // doubleClick={{ disabled: true }} // optional: disable double-click zoom
-          >
+        <div style={{ display: "flex", alignItems: "center", backgroundColor: "black" }}>
+          <TransformWrapper defaultScale={1} wheel={{ step: 0.4 }} pinch={{ step: 0.2 }}>
             <TransformComponent>
               <div style={{ position: "relative", width: "100%" }}>
                 <Image
                   src={currentImage.publicURL}
                   fit="contain"
-                  // height="100vh"
-                  // width="90vw"
                   alt="Enlarged wildlife image"
                 />
                 {showAIBoxes &&
@@ -681,6 +359,7 @@ export function ImageAnnotation({ fetchNextImage, filters }) {
           </TransformWrapper>
         </div>
       </Modal>
+
       <Modal
         opened={showVideo}
         onClose={() => setShowVideo(false)}
@@ -688,12 +367,7 @@ export function ImageAnnotation({ fetchNextImage, filters }) {
         size="lg"
       >
         {currentImage?.videoUrl && (
-          <video
-            src={currentImage.videoUrl}
-            controls
-            autoPlay
-            style={{ width: "100%", maxHeight: "80vh" }}
-          >
+          <video src={currentImage.videoUrl} controls autoPlay style={{ width: "100%", maxHeight: "80vh" }}>
             Your browser does not support the video tag.
           </video>
         )}
